- Downgrade net core sdk
- Merge float and int converters
- Remove unneeded allocations
