+++ global.json
@@ -1,6 +1,6 @@
 {
-    "projects": [ "src", "tests" ],
-    "sdk": {
-       "version": "1.0.0-preview4-004233" 
-    }
+  "projects": ["src", "tests"],
+  "sdk": {
+    "version": "1.0.0-rc3-004530"
+  }
 }
