+++ src/msgpack.light/ExceptionUtils.cs
@@ -45,12 +45,12 @@ namespace ProGaudi.MsgPack.Light
             return new SerializationException($"Waited for an int, got {type:G} (0x{type:X})");
         }
 
-        public static Exception IntSerializationFailture(long value)
+        public static Exception IntSerializationFailure(long value)
         {
             return new SerializationException($"Can't serialize {value}");
         }
 
-        public static Exception IntSerializationFailture(ulong value)
+        public static Exception IntSerializationFailure(ulong value)
         {
             return new SerializationException($"Can't serialize {value}");
         }
