+++ src/msgpack.light/MsgPackContext.cs
@@ -16,16 +16,16 @@ namespace ProGaudi.MsgPack.Light
             {typeof (bool), new BoolConverter()},
             {typeof (string), new StringConverter()},
             {typeof (byte[]), new BinaryConverter()},
-            {typeof (float), new FloatConverter()},
-            {typeof (double), new FloatConverter()},
-            {typeof (byte), new IntConverter()},
-            {typeof (sbyte), new IntConverter()},
-            {typeof (short), new IntConverter()},
-            {typeof (ushort), new IntConverter()},
-            {typeof (int), new IntConverter()},
-            {typeof (uint), new IntConverter()},
-            {typeof (long), new IntConverter()},
-            {typeof (ulong), new IntConverter()},
+            {typeof (float), new NumberConverter()},
+            {typeof (double), new NumberConverter()},
+            {typeof (byte), new NumberConverter()},
+            {typeof (sbyte), new NumberConverter()},
+            {typeof (short), new NumberConverter()},
+            {typeof (ushort), new NumberConverter()},
+            {typeof (int), new NumberConverter()},
+            {typeof (uint), new NumberConverter()},
+            {typeof (long), new NumberConverter()},
+            {typeof (ulong), new NumberConverter()},
             {typeof (DateTime), new DateTimeConverter()},
             {typeof (DateTimeOffset), new DateTimeConverter()},
             {typeof (TimeSpan), new TimeSpanConverter() },
